+++ src/main/java/Duke/Duke.java
@@ -1,0 +1,33 @@
+package Duke;
+
+/**
+ * Duke.Duke is a simple bot that allows users to keep track of different types of tasks.
+ */
+public class Duke {
+    private final Storage storage;
+    private final TaskList tasks;
+    private final UI ui;
+
+    public Duke(String filePath) {
+        this.storage = new Storage(filePath);
+        this.tasks = this.storage.loadTasks();
+        this.ui = new UI();
+    }
+
+    public void run() {
+        UI.greet();
+        Parser parser = new Parser(this.tasks);
+        while (!parser.isExit()) {
+            String userInput = ui.readInput();
+            parser.commands(userInput);
+            storage.saveData(tasks.encodeTasks());
+        }
+    }
+
+    public static void main(String[] args)  {
+        new Duke("data/Duke.Duke.txt").run();
+
+    }
+}
+
+
