+++ src/main/java/Duke/Exception/InvalidCommandException.java
@@ -1,11 +1,13 @@
+package Duke.Exception;
+
 /**
- * InvalidCommandException is an exception that is thrown when command is not recognized for any type of task.
+ * Duke.Exception.InvalidCommandException is an exception that is thrown when command is not recognized for any type of task.
  */
 public class InvalidCommandException extends Exception {
     private final static String SAD_ROBOT_ICON = "[~T-T~]";
 
     /**
-     * Constructs error message for InvalidCommandException.
+     * Constructs error message for Duke.Exception.InvalidCommandException.
      */
     public InvalidCommandException() {
         super(SAD_ROBOT_ICON +": OOPS!!! I'm sorry, but I don't know what that means.");
