+++ src/main/java/Duke/Exception/InvalidIndexException.java
@@ -1,5 +1,7 @@
+package Duke.Exception;
+
 /**
- * InvalidIndexException is thrown if done i, where i is the index of task is not a number or taskNumber
+ * Duke.Exception.InvalidIndexException is thrown if done i, where i is the index of task is not a number or taskNumber
  * exceeds the current list size.
  */
 public class InvalidIndexException extends Exception {
